Add bulk cart functionality and user existence check
The source code has been updated to add a new feature to get all carts of a user by userId in the `cart.controller.ts` file. A new DTO `CartBulkDTO` was imported and used in the new function `getAllUserCarts()`. New codes were necessary to ensure users can now view all their carts in bulk rather than individually, improving UX.

In the `auth.controller.ts` file, a validation was added to the `register` method to check the existence of a user before registration. This prevents the creation of a user with a username already in use, hence avoiding potential clashes.